updated install code and beginnings of update code
